+++ release_notes.md
@@ -1,13 +1,29 @@
 DFTools Changelog
 =================
 
+## Release version 1.6.0
+*2018.05.14*
+
+### New Feature
+
+### Changes
+* Adding vertex end reference to dag init vertices; 
+* Setting the INIT_SIZE property of a DAGInitVertex to the same value of a SDFInitVertex;
+* Copy all properties of an SDFVertex to a DAGVertex;
+* Retrieving Sink and Source port name of SDFVertex to DAGVertex;
+* Using DAGEdge in the aggregate instead of SDFEdge now;
+
+### Bug fix
+* Force activation of UI plugin to have the properties tested properly;
+
 ## Release version 1.5.0
 *2018.04.24*
 
 ### New Feature
 
 ### Changes
-* Add properties to the DAG vertices;
+* Add data size property to SDF edges;
+* Use the data size property when computing DAG edges weight in DAG transformation;
 
 ### Bug fix